Replaced memory pointer in getBalance() with storage pointer - partially addresses #15
The getBalance() function in IPool.sol now uses a storage pointer instead of a memory pointer. Testing showed that using the storage pointer decreased the cost of getBalance by about 500 gas, but made getUsedBalance and getMinimumBalance more expensive. This is likely because the other two functions read an extra field from the record.