Merged IndexPoolMeta and PoolUpdateRecord structs and fixed errors with the way they were used. | Resolves #5 #7 #10 #13
Made all of these changes as a single PR since they are related to the use of the structs which were modified.

Fixed erroneous typecast to uint8 for categoryID in . #5

Merged IndexPoolMeta and PoolUpdateRecord into a single struct. #13

Replaced  function with modifier of the same name. Replaced references in all functions that used it except for , which already has access to the struct. #7

Set pool's last reweigh timestamp in  #10

Modified order of execution to apply storage updates to meta prior to calling reweigh and reindex. This is unlikely to be an issue due to the fact that the pool is always an approved implementation, but there's no efficiency impact so to stay on the safe side I switched the order.